fix argument mismatch with solve_periodic_interp_1d_d
git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@5602 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/einspline/bspline_create.c
@@ -1008,7 +1008,7 @@ solve_deriv_interp_1d_d (double bands[], double coefs[],
 // On exit, coefs with contain interpolating B-spline coefs
 void 
 solve_periodic_interp_1d_d (double bands[], double coefs[],
-			    int M, int cstride)
+			    int M, intptr_t cstride)
 {
   double lastCol[M];
   // Now solve:
